+++ TheMountain/Vertex.h
@@ -3,12 +3,13 @@
 class Vertex
 {
 private:
-	int x, y, z;
+	float x, y, z;
 public:
 	Vertex()=default;
-	Vertex(int, int, int);
-	int GetX();
-	int GetY();
-	int GetZ();
+	Vertex(float, float, float);
+	float GetX();
+	float GetY();
+	float GetZ();
 	friend std::ostream& operator<<(std::ostream& out, Vertex& obj);
+	//friend Vector operator-(const Vertex&, const Vertex&);
 };
